+++ MixItUp.Base/ViewModel/MainControls/CommunityCommandsMainControlViewModel.cs
@@ -13,13 +13,6 @@ namespace MixItUp.Base.ViewModel.MainControls
     {
         private const int SearchResultsPageSize = 25;
 
-        /// <summary>
-        /// Set to true to disable all backend calls during infrastructure migration.
-        /// Set to false when the new backend is ready.
-        /// In the future, this will be an API call in MixItUpService.cs to enable/disable community commands from the dashboard
-        /// </summary>
-        public const bool IsMaintenanceMode = true;
-
         public ICommand BackCommand { get; set; }
 
         public bool ShowHome
@@ -228,7 +221,7 @@ namespace MixItUp.Base.ViewModel.MainControls
                 }
                 catch (Exception ex)
                 {
-                    Logger.Log(ex);
+                    await this.HandleCommunityCommandsException(ex);
                 }
             });
 
@@ -242,7 +235,7 @@ namespace MixItUp.Base.ViewModel.MainControls
                 }
                 catch (Exception ex)
                 {
-                    Logger.Log(ex);
+                    await this.HandleCommunityCommandsException(ex);
                 }
             });
 
@@ -261,7 +254,7 @@ namespace MixItUp.Base.ViewModel.MainControls
                 }
                 catch (Exception ex)
                 {
-                    Logger.Log(ex);
+                    await this.HandleCommunityCommandsException(ex);
                 }
             });
 
@@ -279,7 +272,7 @@ namespace MixItUp.Base.ViewModel.MainControls
                 }
                 catch (Exception ex)
                 {
-                    Logger.Log(ex);
+                    await this.HandleCommunityCommandsException(ex);
                 }
             });
 
@@ -292,7 +285,7 @@ namespace MixItUp.Base.ViewModel.MainControls
                 }
                 catch (Exception ex)
                 {
-                    Logger.Log(ex);
+                    await this.HandleCommunityCommandsException(ex);
                 }
             });
 
@@ -311,7 +304,7 @@ namespace MixItUp.Base.ViewModel.MainControls
                 }
                 catch (Exception ex)
                 {
-                    Logger.Log(ex);
+                    await this.HandleCommunityCommandsException(ex);
                 }
             });
 
@@ -328,75 +321,96 @@ namespace MixItUp.Base.ViewModel.MainControls
                 }
                 catch (Exception ex)
                 {
-                    Logger.Log(ex);
+                    await this.HandleCommunityCommandsException(ex);
                 }
             });
 
             this.PreviousResultsCommand = this.CreateCommand(async () =>
             {
-                this.CurrentResultsPage--;
-                if (this.ShowMyCommands)
+                try
                 {
-                    await this.PerformMyCommands();
+                    this.CurrentResultsPage--;
+                    if (this.ShowMyCommands)
+                    {
+                        await this.PerformMyCommands();
+                    }
+                    else if (this.ShowUserCommands)
+                    {
+                        await this.PerformUserCommands();
+                    }
+                    else if (this.ShowSearch)
+                    {
+                        await this.PerformSearch();
+                    }
                 }
-                else if (this.ShowUserCommands)
+                catch (Exception ex)
                 {
-                    await this.PerformUserCommands();
-                }
-                else if (this.ShowSearch)
-                {
-                    await this.PerformSearch();
+                    await this.HandleCommunityCommandsException(ex);
                 }
             });
 
             this.NextResultsCommand = this.CreateCommand(async () =>
             {
-                this.CurrentResultsPage++;
-                if (this.ShowMyCommands)
+                try
                 {
-                    await this.PerformMyCommands();
+                    this.CurrentResultsPage++;
+                    if (this.ShowMyCommands)
+                    {
+                        await this.PerformMyCommands();
+                    }
+                    else if (this.ShowUserCommands)
+                    {
+                        await this.PerformUserCommands();
+                    }
+                    else if (this.ShowSearch)
+                    {
+                        await this.PerformSearch();
+                    }
                 }
-                else if (this.ShowUserCommands)
+                catch (Exception ex)
                 {
-                    await this.PerformUserCommands();
-                }
-                else if (this.ShowSearch)
-                {
-                    await this.PerformSearch();
+                    await this.HandleCommunityCommandsException(ex);
                 }
             });
         }
 
         public async Task ReviewCommand(int rating, string review)
         {
-            await ServiceManager.Get<MixItUpService>().AddReview(new CommunityCommandReviewModel()
+            try
             {
-                CommandID = this.CommandDetails.ID,
-                Rating = rating,
-                Review = review,
-            });
+                await ServiceManager.Get<MixItUpService>().AddReview(new CommunityCommandReviewModel()
+                {
+                    CommandID = this.CommandDetails.ID,
+                    Rating = rating,
+                    Review = review,
+                });
 
-            this.GetCommandDetailsCommand.Execute(this.CommandDetails.ID);
+                this.GetCommandDetailsCommand.Execute(this.CommandDetails.ID);
+            }
+            catch (Exception ex)
+            {
+                await this.HandleCommunityCommandsException(ex);
+            }
         }
 
         public async Task ReportCommand(string report)
         {
-            await ServiceManager.Get<MixItUpService>().ReportCommand(new CommunityCommandReportModel()
+            try
             {
-                CommandID = this.CommandDetails.ID,
-                Report = report
-            });
+                await ServiceManager.Get<MixItUpService>().ReportCommand(new CommunityCommandReportModel()
+                {
+                    CommandID = this.CommandDetails.ID,
+                    Report = report
+                });
+            }
+            catch (Exception ex)
+            {
+                await this.HandleCommunityCommandsException(ex);
+            }
         }
 
         protected override async Task OnVisibleInternal()
         {
-            // Skip all backend calls during maintenance mode
-            if (IsMaintenanceMode)
-            {
-                await base.OnVisibleInternal();
-                return;
-            }
-
             if (this.firstLoadCompleted)
             {
                 if (this.ShowHome)
@@ -410,12 +424,6 @@ namespace MixItUp.Base.ViewModel.MainControls
 
         private async Task NavigateToCategories()
         {
-            // Skip all backend calls during maintenance mode
-            if (IsMaintenanceMode)
-            {
-                return;
-            }
-
             if (this.lastCategoryRefresh.TotalMinutesFromNow() > 1)
             {
                 try
@@ -431,7 +439,7 @@ namespace MixItUp.Base.ViewModel.MainControls
                 }
                 catch (Exception ex)
                 {
-                    Logger.Log(ex);
+                    await this.HandleCommunityCommandsException(ex);
                 }
 
                 this.lastCategoryRefresh = DateTimeOffset.Now;
@@ -509,6 +517,18 @@ namespace MixItUp.Base.ViewModel.MainControls
             this.ShowCommandDetails = false;
             this.ShowUserCommands = false;
             this.ShowMyCommands = false;
+        }
+
+        private async Task HandleCommunityCommandsException(Exception ex)
+        {
+            if (ex is CommunityCommandsUnavailableException)
+            {
+                await DialogHelper.ShowMessage(ex.Message);
+            }
+            else
+            {
+                Logger.Log(ex);
+            }
         }
     }
 }